[FAB-2036] auto-vendor external dependencies
We add a feature to automatically include external dependencies
in the CAR payload during 'chaintool package' operations.  This
is similar to what CR 6991 did for the peer CLI.

As before, we try to optimize/minimize the payload by excluding
things that we can safely assume that the peer will provide.  This
includes the GOROOT environment, the chaincode shim, and any files
that chaintool itself generates, since they will be regenerated
by chaintool on the peer side.

Fixes FAB-2036

Change-Id: I0f2e633447b2650bedefd9d81fb539eae105cbbb
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>