[FAB-9327] Support 'build -p' for system chaincode
The 'chaintool build' for system-chaincodes currently has a limitation
where it only supports being run from the CWD of the chaincode.  This
means that the "-p" option cannot be used.

This limitation is due to the execution of the "go list" command as
part of discovering context.  This command must be run in the CWD of
the package we are discovering.  We can enhance chaintool by managing
the CWD while we launch this command and once again support the -p
switch.  This will be important later one when we add broader support
for SCC as part of FAB-9191/FAB-8724.

Fixes FAB-9327

Change-Id: If7f791301df347749ba26f2824929896e2020657
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>