[FAB-10209] subcommand proto multi-file support
This changes proto subcmd to accept a list of args or file glob descriptor as input to chaintool proto.

This change eases conversion of .cci to proto, allowing client .proto import.

Changes made to:
  chaintool.core/subcommand-descriptors
  chaintool.subcommands.proto/run
allow for collections of file names

Changes made to:
  chaintool.subcommands.proto/getoutputfile
support --output option being a directory

All tests in the project pass

Change-Id: Ic8821dce4f032c3fd476af4ac2dded09209573ef
Signed-off-by: Matt Rkiouak <rkiouak@me.com>